Better code organization for GetUserActivities
